Move io models to io module
